Call the GetClient method to retrieve an HttpClient object to work with. Remember that GetClient is asynchronous, so use the await operator to capture the object returned by this method.
Call the GetStringAsync method of the HttpClient object and provide the base URL to retrieve an array of parts from the REST web service. The data is returned asynchronously as a JSON string.

Deserialize the JSON string returned by this method into a list of Part objects using the JsonSerializer.Deserialize method. Return this list to the caller.

The completed method should look like this: