Finally, return the new inserted Part.
Build and run the app. Select the Add New Part button and enter a name, type, and supplier to create a new part. Select Save. The Add method in the PartsManager class will get invoked, which creates the new part in the web service. If the operation is successful, the part list page will reappear with the new part at the bottom of the list.